Fix cli tests
