+++ raiden/tests/integration/cli/conftest.py
@@ -1,22 +1,46 @@
+import os
+import sys
+from copy import copy
+
+import pexpect
 import pytest
 
 from raiden.settings import RED_EYES_CONTRACT_VERSION
-from raiden.tests.utils.smoketest import setup_testchain_and_raiden
+from raiden.tests.utils.smoketest import setup_raiden, setup_testchain
 
 
 @pytest.fixture(scope='session')
-def blockchain_provider():
-    result = setup_testchain_and_raiden(
+def testchain_provider():
+    testchain = setup_testchain(print_step=lambda x: None)
+
+    yield testchain
+
+    for geth_process in testchain['processes_list']:
+        geth_process.kill()
+
+
+@pytest.fixture(scope='module')
+def cli_tests_contracts_version():
+    return RED_EYES_CONTRACT_VERSION
+
+
+@pytest.fixture(scope='module')
+def raiden_testchain(testchain_provider, cli_tests_contracts_version):
+    import time
+    start_time = time.monotonic()
+
+    result = setup_raiden(
         transport='matrix',
         matrix_server='auto',
         print_step=lambda x: None,
-        # cli tests should work with production contracts
-        contracts_version=RED_EYES_CONTRACT_VERSION,
+        contracts_version=cli_tests_contracts_version,
+        testchain_setup=testchain_provider,
     )
     args = result['args']
     # The setup of the testchain returns a TextIOWrapper but
     # for the tests we need a filename
     args['password_file'] = args['password_file'].name
+    print('setup_raiden took', time.monotonic() - start_time)
     return args
 
 
@@ -31,8 +55,8 @@ def changed_args():
 
 
 @pytest.fixture()
-def cli_args(blockchain_provider, removed_args, changed_args):
-    initial_args = blockchain_provider.copy()
+def cli_args(raiden_testchain, removed_args, changed_args):
+    initial_args = raiden_testchain.copy()
 
     if removed_args is not None:
         for arg in removed_args:
@@ -64,3 +88,19 @@ def cli_args(blockchain_provider, removed_args, changed_args):
                 args.append(arg_value)
 
     return args
+
+
+@pytest.fixture
+def raiden_spawner(tmp_path):
+    def spawn_raiden(args):
+        # Remove any possibly defined `RAIDEN_*` environment variables from outer scope
+        new_env = {k: copy(v) for k, v in os.environ.items() if not k.startswith('RAIDEN')}
+        new_env['HOME'] = str(tmp_path)
+
+        return pexpect.spawn(
+            sys.executable, ['-m', 'raiden'] + args,
+            logfile=sys.stdout,
+            encoding='utf-8',
+            env=new_env,
+        )
+    return spawn_raiden
