+++ raiden/ui/cli.py
@@ -69,7 +69,7 @@ def options(func):
         option(
             '--datadir',
             help='Directory for storing raiden data.',
-            default=os.path.join(os.path.expanduser('~'), '.raiden'),
+            default=lambda: os.path.join(os.path.expanduser('~'), '.raiden'),
             type=click.Path(
                 exists=False,
                 dir_okay=True,
